Migrate separator from /docs/components/separator to /components/separator
Port all examples (Horizontal, Vertical, Menu, List) from the old docs page
to the new reference page, remove the old page, and update all navigation
links, routes, and E2E tests to use the new path.
