+++ site/ui/routes.tsx
@@ -60,7 +60,6 @@ import { ResizablePage } from './pages/resizable'
 import { ResizableRefPage } from './pages/components/resizable'
 import { ScrollAreaPage } from './pages/scroll-area'
 import { ScrollAreaRefPage } from './pages/components/scroll-area'
-import { SeparatorPage } from './pages/separator'
 import { SeparatorRefPage } from './pages/components/separator'
 import { SkeletonPage } from './pages/skeleton'
 import { TextareaPage } from './pages/textarea'
@@ -258,7 +257,7 @@ export function createApp() {
               <h3 className="text-sm font-medium text-foreground group-hover:text-foreground">Sidebar</h3>
               <p className="text-xs text-muted-foreground">Collapsible navigation panel</p>
             </a>
-            <a href="/docs/components/separator" className="group flex flex-col rounded-xl border border-border hover:border-ring transition-colors no-underline p-6 space-y-2">
+            <a href="/components/separator" className="group flex flex-col rounded-xl border border-border hover:border-ring transition-colors no-underline p-6 space-y-2">
               <h3 className="text-sm font-medium text-foreground group-hover:text-foreground">Separator</h3>
               <p className="text-xs text-muted-foreground">Visual divider between content</p>
             </a>
@@ -603,12 +602,7 @@ export function createApp() {
     return c.render(<SelectPage />)
   })
 
-  // Separator documentation
-  app.get('/docs/components/separator', (c) => {
-    return c.render(<SeparatorPage />)
-  })
-
-  // Separator reference page (redesigned #515)
+  // Separator reference page
   app.get('/components/separator', (c) => {
     return c.render(<SeparatorRefPage />)
   })
